fix(vertexai): Don't check for unused api_key.
- Split `test_google_provider` into `test_gemini_provider` and `test_vertexai_provider`.
