+++ tests/unit/providers/test_vertexai_provider.py
@@ -1,0 +1,8 @@
+from any_llm.config import ClientConfig
+from any_llm.providers.vertexai import VertexaiProvider
+
+
+def test_vertexai_initialization_without_api_key() -> None:
+    """Test that the VertexaiProvider initializes correctly without API Key."""
+    provider = VertexaiProvider(ClientConfig())
+    assert provider.config.api_key is None
